backend/requirements.txt
fastapi>=0.115.6,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
python-multipart>=0.0.20,<1.0.0
pydantic>=2.10.4,<3.0.0
openai>=2.26.0,<3.0.0
chromadb>=1.5.5,<2.0.0
dashscope>=1.23.0,<2.0.0
langchain>=1.2.14,<2.0.0
langchain-core>=1.2.21,<2.0.0
langchain-community>=0.4.1,<1.0.0
PyMuPDF>=1.25.1,<2.0.0
python-docx>=1.1.2,<2.0.0
